pyproject.toml
[tool.poetry]
name = "detraffic"
version = "0.1.0"
description = ""
authors = ["Melih DARCAN <57872471+MelihDarcanxyz@users.noreply.github.com>"]
license = "License :: OSI Approved :: GNU General Public License v3 (GPLv3)"

[tool.poetry.dependencies]
python = "^3.10"
sumo-rl = "^1.4.5"
pettingzoo = "^1.24.3"
torch = "^2.2.2"
matplotlib = "^3.8.4"
tqdm = "^4.66.4"
PyYAML = "^6.0.1"

[tool.poetry.dev-dependencies]
pdoc = "^14.4.0"
pytest = "^8.1.0"
mypy = "^1.8.0"
pre-commit = "^3.6.2"
ruff = "^0.3.1"
refurb = "^1.28.0"
ipykernel = "^6.29.4"
ipywidgets = "^8.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Ruff linting and formatting
[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
extend-select = ["I", "D", "UP"] # I for isort, D for pydocstyle, UP for pyupdate
extend-ignore = ["D203", "D212"]

# Mypy static type checking
[tool.mypy]
python_version = "3.10"
